% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ahringer}
\alias{ahringer}
\title{Ahringer RNAi library annotations
The Caenorhabditis elegans RNAi feeding library was constructed by Julie
Ahringer's group at The Wellcome CRC Institute, University of Cambridge,
Cambridge, UK and is distributed by Source BioScience. It is designed for
genome wide study of gene function in C. elegans through loss of function
studies. It can be used both for rapid large scale RNAi experiments and for
the study of individual genes.}
\format{A data frame with 13 variables:
  \describe{
  \item{\code{cloneId}}{a character vector}
  \item{\code{plate}}{a character vector}
  \item{\code{well}}{a character vector}
  \item{\code{chrom}}{a character vector}
  \item{\code{orfOriginal}}{a character vector}
  \item{\code{fwdPrimerSeq}}{a character vector}
  \item{\code{revPrimerSeq}}{a character vector}
  \item{\code{sourceBioscienceLocation}}{a character vector}
  \item{\code{extraInfo}}{a character vector}
  \item{\code{geneId}}{a character vector}
  \item{\code{orf}}{a character vector}
  \item{\code{publicName}}{a character vector}
  \item{\code{wormbaseGeneOtherIds}}{a character vector}
  }}
\source{
Source BioScience (\url{http://www.us.lifesciences.sourcebioscience.com/clone-production/non-mammalian/c-elegans/c-elegans-rnai-library})
}
\usage{
ahringer
}
\description{
Ahringer RNAi library annotations
The Caenorhabditis elegans RNAi feeding library was constructed by Julie
Ahringer's group at The Wellcome CRC Institute, University of Cambridge,
Cambridge, UK and is distributed by Source BioScience. It is designed for
genome wide study of gene function in C. elegans through loss of function
studies. It can be used both for rapid large scale RNAi experiments and for
the study of individual genes.
}
\keyword{datasets}

