% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationFile.R
\name{annotationFile}
\alias{annotationFile}
\title{Annotation File}
\usage{
annotationFile(file, version = NULL, dir = ".")
}
\arguments{
\item{file}{Annotation file name.}

\item{version}{WormBase version. If \code{NULL}, defaults to current production
release version available on the WormBase website. Legacy versions can
be specified as a string (e.g. "WS262").}

\item{dir}{Output directory.}
}
\value{
Invisible file path vector.
}
\description{
Annotation File
}
\examples{
# Current
current <- annotationFile("geneIDs")
print(current)

# Versioned
versioned <- annotationFile("geneIDs", version = "WS262")
print(versioned)

# Clean up
file_exists(c(current, versioned))
file_delete(c(current, versioned))
}
\seealso{
Other Annotation File Functions: \code{\link{blastp}},
  \code{\link{description}}, \code{\link{geneIDs}},
  \code{\link{geneOtherIDs}}, \code{\link{oligos}},
  \code{\link{orthologs}}
}
\keyword{internal}
