% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotations}
\alias{annotations}
\title{Gene annotations}
\format{An object of class \code{list} of length 6.}
\usage{
annotations
}
\description{
WormBase, Ensembl, PANTHER, and R version information.
}
\keyword{datasets}
