% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worminfo.R
\docType{package}
\name{bcbioRnaseq}
\alias{bcbioRnaseq}
\alias{bcbioRnaseq-package}
\title{\code{worminfo} package}
\description{
C. elegans genome annotations and RNAi clone mappings.
}
\details{
See the README on \href{https://github.com/steinbaugh.com/worminfo}{GitHub}.
}
