% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bindMetadata}
\alias{bindMetadata}
\title{Bind metadata to data.frame}
\usage{
bindMetadata(df, id = "geneId", type = "simple")
}
\arguments{
\item{df}{data.frame with identifier column}

\item{id}{Identifier type (geneID, orf, publicName)}

\item{type}{Output type (report, simple)}
}
\value{
data.frame cbind with metadata
}
\description{
Bind metadata to data.frame
}

