% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bindOrfMetadata}
\alias{bindOrfMetadata}
\title{Bind ORF metadata to data.frame}
\usage{
bindOrfMetadata(df)
}
\arguments{
\item{df}{data.frame with `orf` column}
}
\value{
data.frame with metadataOrf information
}
\description{
Bind ORF metadata to data.frame
}

