% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-blastp.R
\name{blastp}
\alias{blastp}
\title{Best BLASTP hits}
\usage{
blastp(version = getOption(x = "wormbase.version", default = NULL))
}
\arguments{
\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS267").}
}
\value{
\code{SplitDataFrameList}.
Split by \code{wormpep} column.
}
\description{
Best BLASTP hits
}
\note{
Updated 2021-02-18.
}
\examples{
x <- blastp()
print(x)
}
