% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bm2wb}
\alias{bm2wb}
\title{Match biomaRt output with WormBase Gene IDs}
\usage{
bm2wb(df)
}
\arguments{
\item{df}{}
}
\value{
data.frame with all WormBase Gene IDs
}
\description{
Match biomaRt output with WormBase Gene IDs
}

