% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{build}
\alias{build}
\title{Build version
WormBase, Ensembl, PANTHER, and R version information}
\format{list}
\usage{
build
}
\description{
Build version
WormBase, Ensembl, PANTHER, and R version information
}
\keyword{datasets}

