% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.R
\name{clone}
\alias{clone}
\title{Feeding RNAi Library clone matching.}
\usage{
clone(id = NULL, library = "orfeome", wells = NULL, select = "simple")
}
\arguments{
\item{id}{Clone identifier.}

\item{library}{Library type ("orfeome" or "ahringer").}

\item{wells}{Library plate format (96, 384).}

\item{select}{Select columns (report, simple).}
}
\value{
tibble with \code{gene()} metadata.
}
\description{
Feeding RNAi Library clone matching.
}
\examples{
clone("11010-G06", library = "orfeome")
clone("11010-G06", library = "orfeome", select = "report")
clone("086-B01", library = "ahringer", wells = 96)
clone("III-006-C01", library = "ahringer", wells = 384)
}

