% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deadOrf.R
\name{deadOrf}
\alias{deadOrf}
\title{Map dead ORF to WormBase Gene ID}
\usage{
deadOrf(orf)
}
\arguments{
\item{orf}{ORF vector.}
}
\value{
tibble
}
\description{
Map dead ORF to WormBase Gene ID
}
\examples{
deadOrf("M01E10.2")
}

