% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deadSequence.R
\name{deadSequence}
\alias{deadSequence}
\title{Map dead ORF to WormBase Gene ID.}
\usage{
deadSequence(sequence)
}
\arguments{
\item{sequence}{Sequence (ORF).}
}
\value{
tibble.
}
\description{
Map dead ORF to WormBase Gene ID.
}
\examples{
deadSequence("M01E10.2")
}

