% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\docType{data}
\name{defaultCol}
\alias{defaultCol}
\title{Default columns for \code{select}}
\format{An object of class \code{character} of length 3.}
\usage{
defaultCol
}
\arguments{
\item{defaultCol}{Default columns}
}
\description{
Default columns for \code{select}
}
\keyword{datasets}

