% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{description}
\alias{description}
\title{Gene Functional Descriptions}
\usage{
description(version = NULL, dir = ".")
}
\arguments{
\item{version}{WormBase version. If \code{NULL}, defaults to current production
release version available on the WormBase website. Legacy versions can
be specified as a string (e.g. "WS262").}

\item{dir}{Output directory.}
}
\value{
Gene \link{tibble}.
}
\description{
Gene Functional Descriptions
}
\examples{
# CPU intensive
description() \%>\% glimpse()
}
\seealso{
Other Annotation File Functions: \code{\link{annotationFile}},
  \code{\link{geneIDs}}, \code{\link{geneOtherIDs}},
  \code{\link{oligos}}, \code{\link{orthologs}}
}
