% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggnog.R
\name{eggnog}
\alias{eggnog}
\title{EggNOG annotations}
\usage{
eggnog(query)
}
\arguments{
\item{query}{EggNOG identifier}
}
\value{
tibble
}
\description{
EggNOG annotations
}
\examples{
eggnog(c("ENOG410ZGMS", "KOG3863"))
}

