% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ensembl2wormbase}
\alias{ensembl2wormbase}
\title{Match Ensembl/biomaRt output with WormBase Gene IDs}
\usage{
ensembl2wormbase(df)
}
\arguments{
\item{df}{data.frame output from biomaRt}
}
\value{
data.frame with all WormBase Gene IDs
}
\description{
Match Ensembl/biomaRt output with WormBase Gene IDs
}

