% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalIDs.R
\name{externalIDs}
\alias{externalIDs}
\title{External Identifiers}
\usage{
externalIDs(genes, progress = TRUE)
}
\arguments{
\item{genes}{\code{character}. Gene identifiers.}

\item{progress}{\code{boolean}. Show progress using pbapply package.}
}
\value{
\code{tbl_df}.
}
\description{
External Identifiers
}
\examples{
x <- externalIDs(
    genes = c("WBGene00000912", "WBGene00004804"),
    progress = FALSE
)
glimpse(x)
}
