% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{gene}
\alias{gene}
\title{Gene annotations.}
\usage{
gene(id = NULL, format = "gene", select = "simple")
}
\arguments{
\item{id}{Gene identifier.}

\item{format}{Identifier type (gene, name, sequence).}

\item{select}{Columns to select (report, simple).}
}
\value{
tibble.
}
\description{
Gene annotations.
}
\examples{
gene(id = "WBGene00004804", format = "gene")
gene(id = "skn-1", format = "name")
gene(id = "T19E7.2", format = "sequence")
}

