% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneAnnotation}
\alias{geneAnnotation}
\title{Gene annotation
Raw data for use with the \code{gene()} function}
\format{tibble}
\usage{
geneAnnotation
}
\description{
Gene annotation
Raw data for use with the \code{gene()} function
}
\keyword{datasets}

