% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneBind.R
\name{geneBind}
\alias{geneBind}
\title{Bind gene annotations to input data.frame}
\usage{
geneBind(df, ...)
}
\arguments{
\item{df}{data.frame with identifier column.}

\item{...}{Optional parameters for `gene` function.}
}
\value{
data.frame cbind with metadata
}
\description{
Bind gene annotations to input data.frame
}

