% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClass.R
\name{geneClass}
\alias{geneClass}
\title{Gene mapping by class abbreviation}
\usage{
geneClass(class, ...)
}
\arguments{
\item{class}{Gene class abbreviation (e.g. "daf")}
}
\description{
Gene mapping by class abbreviation
}
\examples{
geneClass(c("daf", "nhr"))
}

