% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExternal.R
\name{geneExternal}
\alias{geneExternal}
\title{Gene External RESTful Query}
\usage{
geneExternal(gene)
}
\arguments{
\item{gene}{Gene identifier.}
}
\value{
\link{tibble}.
}
\description{
Gene External RESTful Query
}
\examples{
geneExternal(c("WBGene00000912", "WBGene00004804")) \%>\% glimpse()
}
