% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneID}
\alias{GeneID}
\title{GeneID}
\format{A data frame with 5 variables: \code{GeneID}, \code{public.name},
\code{ORF}, \code{wormbase.status} and \code{gene.other.ids}.}
\usage{
GeneID
}
\description{
A dataset containing all current WormBase GeneIDs, public names, ORFs,
WormBase status, and previously used IDs.
}
\keyword{datasets}

