% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneId2entrezId.R
\name{geneId2entrezId}
\alias{geneId2entrezId}
\title{Get Entrez identifier from WormBase Gene identifier.}
\usage{
geneId2entrezId(geneId)
}
\arguments{
\item{geneId}{WormBase gene identifier vector.}
}
\value{
tibble
}
\description{
Get Entrez identifier from WormBase Gene identifier.
}
\examples{
geneId2entrezId("WBGene00000001")
}

