\name{GeneID_vec}
\alias{GeneID_vec}
\docType{data}
\title{WormBase GeneID vector}
\description{
A list of all the current WormBase GeneIDs available. Useful for setting data
frames to the same number of rows.
}
\usage{data("GeneID_vec")}
\format{
A character vector with 50930 observations.
}
\source{
WormBase (\url{http://www.wormbase.org})
}
\examples{
data(GeneID_vec)
}
\keyword{datasets}
