% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geneIdRows}
\alias{geneIdRows}
\title{Set number of rows to Wormbase Gene IDs}
\usage{
geneIdRows(df)
}
\arguments{
\item{df}{data.frame without all WormBase identifiers}
}
\value{
data.frame with all WormBase identifiers
}
\description{
Set number of rows to Wormbase Gene IDs
}

