% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneIDs.R
\name{geneIDs}
\alias{geneIDs}
\title{Gene Identifiers}
\usage{
geneIDs(version = NULL, dir = ".")
}
\arguments{
\item{version}{WormBase version. If \code{NULL}, defaults to current production
release version available on the WormBase website. Legacy versions can
be specified as a string (e.g. "WS262").}

\item{dir}{Output directory.}
}
\value{
Gene \link{tibble}.
}
\description{
Gene Identifiers
}
\examples{
geneIDs() \%>\% glimpse()
}
\seealso{
Other Annotation File Functions: \code{\link{annotationFile}},
  \code{\link{description}}, \code{\link{geneOtherIDs}},
  \code{\link{oligos}}, \code{\link{orthologs}}
}
