% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntology.R
\name{geneOntology}
\alias{geneOntology}
\title{Gene Ontology RESTful Query}
\usage{
geneOntology(gene)
}
\arguments{
\item{gene}{Gene identifier.}
}
\value{
\link{tibble}.
}
\description{
Gene Ontology RESTful Query
}
\examples{
geneOntology(c("WBGene00000912", "WBGene00004804")) \%>\% glimpse()
}
