% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOtherIDs.R
\name{geneOtherIDs}
\alias{geneOtherIDs}
\title{Other gene identifiers}
\usage{
geneOtherIDs(version = NULL, dir = ".")
}
\arguments{
\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS262").}

\item{dir}{\code{character(1)}.
Output directory.}
}
\value{
\code{tbl_df}.
}
\description{
Other gene identifiers
}
\examples{
x <- geneOtherIDs()
glimpse(x)
}
