% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{general}
\alias{general}
\title{General Arguments}
\arguments{
\item{dir}{Output directory.}

\item{file}{Annotation file name.}

\item{gene}{Gene identifier.}

\item{version}{WormBase version. If \code{NULL}, defaults to current production
release version available on the WormBase website. Legacy versions can
be specified as a string (e.g. "WS262").}
}
\description{
General Arguments
}
