% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getOrfMetadata}
\alias{getOrfMetadata}
\title{ORF metadata matching}
\usage{
getOrfMetadata(orf)
}
\arguments{
\item{orf}{ORF character vector}
}
\value{
data.frame
}
\description{
ORF metadata matching
}

