% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepString.R
\name{grepString}
\alias{grepString}
\title{grep match for toString}
\usage{
grepString(identifier)
}
\arguments{
\item{identifier}{Identifier}
}
\value{
grep \code{string}
}
\description{
grep match for toString
}
\keyword{internal}

