% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical2rnai.R
\name{historical2rnai}
\alias{historical2rnai}
\title{Convert historical RNAi identifier to WBRNAi with WormBase RESTful API.}
\usage{
historical2rnai(historical)
}
\arguments{
\item{historical}{WormBase historical RNAi experiment vector.}
}
\value{
tibble.
}
\description{
Convert historical RNAi identifier to WBRNAi with WormBase RESTful API.
}
\examples{
historical2rnai("JA:K10E9.1")
}

