% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical2wbrnai.R
\name{historical2wbrnai}
\alias{historical2wbrnai}
\title{Convert historical RNAi identifier to WBRNAi with WormBase RESTful API}
\usage{
historical2wbrnai(vec)
}
\arguments{
\item{vec}{Vector of WormBase historical RNAi clones}
}
\value{
data.frame
}
\description{
Convert historical RNAi identifier to WBRNAi with WormBase RESTful API
}
\examples{
historical2wbrnai("JA:K10E9.1", "JA:F56C11.1")
}

