% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{keywordCol}
\alias{keywordCol}
\title{Keyword columns}
\format{An object of class \code{character} of length 14.}
\usage{
keywordCol
}
\arguments{
\item{keywordCol}{Keyword columns}
}
\description{
Keyword columns
}
\keyword{datasets}

