% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{loadData}
\alias{loadData}
\title{Load data and source if necessary}
\usage{
loadData(data)
}
\arguments{
\item{data}{Data filename}
}
\description{
Load data and source if necessary
}
\keyword{internal}

