% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownList.R
\name{markdownList}
\alias{markdownList}
\title{Markdown formatted list}
\usage{
markdownList(vector)
}
\arguments{
\item{vector}{Character vector}
}
\value{
Markdown list
}
\description{
Markdown formatted list
}

