% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{metadata}
\alias{metadata}
\title{Annotation metadata}
\usage{
metadata(rowNames = NULL, id = "geneId", type = "simple")
}
\arguments{
\item{rowNames}{Identifier vector}

\item{id}{Identifier type (geneID, orf, publicName)}

\item{type}{Output type (report, simple)}
}
\value{
metadata data.frame
}
\description{
Annotation metadata
}

