% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_simple}
\alias{metadata_simple}
\title{metadata_simple}
\format{A data frame with 4 variables: \code{GeneID}, \code{ORF},
\code{public.name}, \code{gene.other.ids}}
\usage{
metadata_simple
}
\description{
A concise data frame containing only GeneID, ORF, public name, and other
previously used IDs.
}
\keyword{datasets}

