% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadataOrf}
\alias{metadataOrf}
\title{metadataOrf}
\format{A data frame with 3 variables: \code{geneId}, \code{orf},
  \code{publicName}}
\usage{
metadataOrf
}
\description{
A dataset containing all current WormBase GeneIDs, public names, open reading
frames (ORFs), and previously used IDs. This dataset is filtered to have
unique ORF matches.
}
\keyword{datasets}

