% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadataReport}
\alias{metadataReport}
\title{metadataReport}
\format{A data frame with 12 variables:
  \describe{
  \item{\code{geneId}}{a character vector}
  \item{\code{orf}}{a character vector}
  \item{\code{publicName}}{a character vector}
  \item{\code{wormbaseGeneOtherIds}}{a character vector}
  \item{\code{wormbaseGeneClassDescription}}{a character vector}
  \item{\code{wormbaseConciseDescription}}{a character vector}
  \item{\code{wormbaseBlastpEnsemblGeneName}}{a character vector}
  \item{\code{wormbaseBlastpEnsemblDescription}}{a character vector}
  \item{\code{wormbaseStatus}}{a character vector}
  \item{\code{ensemblGeneBiotype}}{a character vector}
  \item{\code{pantherFamilyName}}{a character vector}
  \item{\code{pantherSubfamilyName}}{a character vector}
  }}
\usage{
metadataReport
}
\description{
A subset of the master metadata file containing only human readable data.
}
\keyword{datasets}

