% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadataSimple}
\alias{metadataSimple}
\title{metadataSimple}
\format{A data frame with 3 variables: \code{geneId}, \code{orf},
  \code{publicName}}
\usage{
metadataSimple
}
\description{
A concise data frame containing only GeneID, ORF, and public name.
}
\keyword{datasets}

