% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microplate.R
\name{microplate}
\alias{microplate}
\title{Generate Well Identifiers for a Microtiter Plate}
\usage{
microplate(plate = 1L, well = 96L, control = 0L, prefix)
}
\arguments{
\item{plate}{Number of plates.}

\item{well}{Number of wells (\code{96}, \code{384}).}

\item{control}{Number of control wells.}

\item{prefix}{\emph{Optional}. Plate name prefix.}
}
\value{
\link{data.frame}.
}
\description{
Generate Well Identifiers for a Microtiter Plate
}
\examples{
microplate() \%>\%
    head()
}
