% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimalClone.R
\name{minimalClone}
\alias{minimalClone}
\title{Minimal RNAi Clones}
\usage{
minimalClone(identifier)
}
\arguments{
\item{identifier}{Clone identifier.}
}
\value{
Character vector.
}
\description{
Strip out unnecessary plate and well separators. Also remove zero padding
for well numbers, for fast and consistent clone matching.
}
\examples{
minimalClone(c("11010-G06", "11010@G06", "III-6-C01"))
}
\keyword{internal}
