% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-oligos.R
\name{oligos}
\alias{oligos}
\title{PCR oligo sequences}
\usage{
oligos(version = getOption(x = "wormbase.version", default = NULL))
}
\arguments{
\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS267").}
}
\value{
\code{DataFrame}.
}
\description{
PCR oligo sequences
}
\note{
Updated 2021-02-17.
}
\examples{
## WormBase FTP server must be accessible.
tryCatch(
    expr = oligos(),
    error = function(e) e
)
}
