% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orfeome.R
\name{orfeome}
\alias{orfeome}
\title{Feeding RNAi Library clone matching}
\usage{
orfeome(id = NULL, library = "orfeome", format = "report")
}
\arguments{
\item{id}{Clone identifier}

\item{library}{Feeding library (ahringer, orfeome)}

\item{format}{Output type (report, simple)}
}
\value{
data.frame with metadata
}
\description{
Feeding RNAi Library clone matching
}
\examples{
rnai(library = "ahringer")
rnai(id = "III-86@B01", library = "ahringer")
rnai(library = "orfeome")
rnai(id = "11010@G06", library = "orfeome")
}

