% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-orthologs.R
\name{orthologs}
\alias{orthologs}
\title{Orthologs}
\usage{
orthologs(version = getOption(x = "wormbase.version", default = NULL),
  progress = FALSE)
}
\arguments{
\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS262").}

\item{progress}{\code{logical(1)}.
Show progress using pbapply package.}
}
\value{
\code{tbl_df}.
}
\description{
Orthologs
}
\note{
Updated 2019-07-27.
}
\examples{
## WormBase FTP server must be accessible.
if (!is.null(curl::nslookup("ftp.wormbase.org"))) {
    x <- orthologs()
    glimpse(x)
}
}
