% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide2gene.R
\name{peptide2gene}
\alias{peptide2gene}
\title{Peptide to Gene}
\usage{
peptide2gene(version = NULL, dir = ".")
}
\arguments{
\item{version}{WormBase version. If \code{NULL}, defaults to current production
release version available on the WormBase website. Legacy versions can
be specified as a string (e.g. "WS262").}

\item{dir}{Output directory.}
}
\value{
\link{tibble}.
}
\description{
Peptide to Gene
}
\examples{
peptide2gene() \%>\% glimpse()
}
