% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clones.R
\name{prettyClone}
\alias{prettyClone}
\title{Pretty RNAi Clones}
\usage{
prettyClone(identifier)
}
\arguments{
\item{identifier}{Clone identifier.}
}
\value{
Character vector.
}
\description{
We strip the formatting variations down to the bare minimum in the annotation
data. This allows for more flexible matching, such as clones that don't
contain a padded well number. At the return step, we reformat the clones back
to a standardized output, separated by dashes.
}
\keyword{internal}
