% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{loadData}
\alias{reexports}
\alias{mdList}
\alias{reexports}
\alias{saveData}
\alias{reexports}
\alias{saveDataRaw}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{\%>\%}}, \code{\link[basejump]{glimpse}}, \code{\link[basejump]{loadData}}, \code{\link[basejump]{mdList}}, \code{\link[basejump]{saveData}}, \code{\link[basejump]{saveDataRaw}}}
}}

