% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{reportCol}
\alias{reportCol}
\title{Report columns}
\format{An object of class \code{character} of length 21.}
\usage{
reportCol
}
\arguments{
\item{reportCol}{Report columns}
}
\description{
Report columns
}
\keyword{datasets}

