% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{rest}
\alias{rest}
\title{WormBase REST API query}
\usage{
rest(url)
}
\arguments{
\item{url}{URL query to WormBase RESTful API}
}
\value{
JSON content
}
\description{
WormBase REST API query
}
\examples{
rest("field/gene/WBGene00000001/gene_class")
}

