% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restGeneOntology.R
\name{restGeneOntology}
\alias{restGeneOntology}
\title{WormBase RESTful RNAi gene ontology query}
\usage{
restGeneOntology(identifier)
}
\arguments{
\item{identifier}{Gene identifier}
}
\value{
JSON content tibble
}
\description{
WormBase RESTful RNAi gene ontology query
}
\examples{
restGeneOntology("WBGene00000001")
}

