% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{restLimit}
\alias{restLimit}
\title{REST query limit}
\format{An object of class \code{numeric} of length 1.}
\usage{
restLimit
}
\arguments{
\item{restLimit}{REST query limit}
}
\description{
REST query limit
}
\keyword{datasets}

