% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{restRnaiReagent}
\alias{restRnaiReagent}
\title{WormBase RESTful RNAi reagent query}
\usage{
restRnaiReagent(rnai)
}
\arguments{
\item{rnai}{RNAi}
}
\value{
JSON content tibble
}
\description{
WormBase RESTful RNAi reagent query
}
\examples{
restRnaiReagent("WBRNAi00000001")
}

