% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restRnaiSequence.R
\name{restRnaiSequence}
\alias{restRnaiSequence}
\title{WormBase RESTful RNAi sequence query}
\usage{
restRnaiSequence(rnai)
}
\arguments{
\item{rnai}{RNAi}
}
\value{
JSON content tibble
}
\description{
WormBase RESTful RNAi sequence query
}
\examples{
restRnaiSequence("WBRNAi00000001")
}

