% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restRnaiTargets.R
\name{restRnaiTargets}
\alias{restRnaiTargets}
\title{WormBase RESTful RNAi targets query}
\usage{
restRnaiTargets(rnai)
}
\arguments{
\item{rnai}{RNAi}
}
\value{
JSON content tibble
}
\description{
WormBase RESTful RNAi targets query
}
\examples{
restRnaiTargets("WBRNAi00000001")
}

