% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnai_orfeome}
\alias{rnai_orfeome}
\title{ORFeome RNAi library annotations}
\format{A data frame with 7 variables:
  \describe{
  \item{\code{ORFeomeID}}{a character vector}
  \item{\code{ORF.original}}{a character vector}
  \item{\code{RNAi.well}}{a character vector}
  \item{\code{GeneID}}{a character vector}
  \item{\code{ORF}}{a character vector}
  \item{\code{public.name}}{a character vector}
  \item{\code{gene.other.ids}}{a character vector}
  }}
\source{
C .elegans RNAi Collection (\url{http://dharmacon.gelifesciences.com/non-mammalian-cdna-and-orf/c.-elegans-rnai}),
WORFDB (\url{http://worfdb.dfci.harvard.edu})
}
\usage{
rnai_orfeome
}
\description{
Developed for RNAi screens in C. elegans, this RNAi collection includes full
open reading frames for over 11,000 genes cloned into a feeding vector.
}
\details{
Derived from the C. elegans ORFeome Library v1.1, the C. elegans RNAi
collection provides comprehensive coverage for screening with over 11,000
RNAi clones. These constructs are cloned into the pL4440-dest-RNAi
Destination vector and are archived as glycerol stocks of the E. coli feeding
strain HT115(DE3). Minipreps of this DNA can be used as templates for in
vitro dsRNA synthesis before RNAi application by soaking or injection.
}
\keyword{datasets}

