% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnaiAnnotation}
\alias{rnaiAnnotation}
\title{RNAi annotation
Raw data for use with the \code{rnai()} function}
\format{tibble}
\usage{
rnaiAnnotation
}
\description{
RNAi annotation
Raw data for use with the \code{rnai()} function
}
\keyword{datasets}

