% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaiPhenotypes.R
\name{rnaiPhenotypes}
\alias{rnaiPhenotypes}
\title{RNAi phenotypes}
\usage{
rnaiPhenotypes(version = NULL, dir = ".", progress = FALSE)
}
\arguments{
\item{version}{\code{string} or \code{NULL}. WormBase version. If \code{NULL} (recommended),
defaults to current production release version available on the WormBase
website. Legacy versions can be specified as a \code{string} (e.g. "WS262").}

\item{dir}{\code{string}. Output directory.}

\item{progress}{\code{boolean}. Show progress using pbapply package.}
}
\value{
\code{tbl_df}.
}
\description{
RNAi phenotypes
}
\examples{
x <- rnaiPhenotypes(progress = FALSE)
glimpse(x)
}
