% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaiReagent.R
\name{rnaiReagent}
\alias{rnaiReagent}
\title{WormBase RESTful RNAi reagent query}
\usage{
rnaiReagent(rnai)
}
\arguments{
\item{rnai}{RNAi}
}
\value{
JSON content tibble
}
\description{
WormBase RESTful RNAi reagent query
}
\examples{
rnaiReagent("WBRNAi00000001")
}

