% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaiSequence.R
\name{rnaiSequence}
\alias{rnaiSequence}
\title{WormBase RESTful RNAi sequence query}
\usage{
rnaiSequence(rnai)
}
\arguments{
\item{rnai}{RNAi}
}
\value{
JSON content tibble
}
\description{
WormBase RESTful RNAi sequence query
}
\examples{
rnaiSequence("WBRNAi00000001")
}

