% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaiTargets.R
\name{rnaiTargets}
\alias{rnaiTargets}
\title{WormBase RESTful RNAi targets query}
\usage{
rnaiTargets(rnai)
}
\arguments{
\item{rnai}{RNAi}
}
\value{
JSON content tibble
}
\description{
WormBase RESTful RNAi targets query
}
\examples{
rnaiTargets("WBRNAi00000001")
}

