% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\docType{data}
\name{ua}
\alias{ua}
\title{User agent for REST API calls}
\format{An object of class \code{character} of length 1.}
\usage{
ua
}
\arguments{
\item{ua}{User agent}
}
\description{
User agent for REST API calls
}
\keyword{internal}

