% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{uniprot}
\alias{uniprot}
\title{UniProt web service query}
\usage{
uniprot(identifier)
}
\arguments{
\item{identifier}{WormBase gene identifier}
}
\value{
tibble
}
\description{
UniProt web service query
}
\author{
Michael Steinbaugh
}
