% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\docType{data}
\name{userAgent}
\alias{userAgent}
\title{User agent for REST API queries}
\format{An object of class \code{character} of length 1.}
\usage{
userAgent
}
\arguments{
\item{userAgent}{User agent}
}
\description{
User agent for REST API queries
}
\keyword{datasets}

