% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worfdb.R
\name{worfdb}
\alias{worfdb}
\alias{worfdbHTML}
\alias{worfdbList}
\alias{worfdbData}
\title{Query WORFDB Website for RNAi Clone Annotations}
\usage{
worfdbHTML(sequence)

worfdbList(worfdbHTML)

worfdbData(worfdbList)
}
\arguments{
\item{sequence}{Sequence identifier.}

\item{worfdbHTML}{List of WORFDB HTML pages.}

\item{worfdbList}{WORFDB list returned by \code{\link[=worfdbList]{worfdbList()}}.}
}
\description{
Query WORFDB Website for RNAi Clone Annotations
}
\keyword{internal}
