% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{wormbase-package}
\alias{wormbase}
\alias{wormbase-package}
\title{wormbase}
\description{
Fetch \emph{Caenorhabditis elegans} genome annotations from WormBase.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wormbase.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/wormbase/}
  \item Report bugs at \url{https://github.com/acidgenomics/wormbase/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
