% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbaseAnnotationFile.R
\name{wormbaseAnnotationFile}
\alias{wormbaseAnnotationFile}
\title{WormBase annotation file}
\usage{
wormbaseAnnotationFile(file)
}
\arguments{
\item{file}{Annotation file (without extension)}
}
\description{
Download an annotation file from the WormBase FTP server
}
\keyword{internal}
