% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbaseFile.R
\name{wormbaseFile}
\alias{wormbaseFile}
\title{Download a file from WormBase}
\usage{
wormbaseFile(file)
}
\arguments{
\item{file}{FTP file request}
}
\value{
File path of download
}
\description{
Download a file from WormBase
}

