% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbase.R
\name{wormbaseGeneMerge}
\alias{wormbaseGeneMerge}
\title{Map a dead sequence to current WormBase gene identifier.}
\usage{
wormbaseGeneMerge(sequence)
}
\arguments{
\item{sequence}{Sequence (ORF).}
}
\value{
tibble.
}
\description{
Map a dead sequence to current WormBase gene identifier.
}
\examples{
wormbaseGeneMerge("M01E10.2")
}

