% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbase.R
\name{wormbaseHistorical2rnai}
\alias{wormbaseHistorical2rnai}
\title{Convert historical RNAi identifier to WBRNAi with WormBase RESTful API.}
\usage{
wormbaseHistorical2rnai(historical)
}
\arguments{
\item{historical}{WormBase historical RNAi experiment vector.}
}
\value{
tibble.
}
\description{
Convert historical RNAi identifier to WBRNAi with WormBase RESTful API.
}
\examples{
wormbaseHistorical2rnai("JA:K10E9.1")
}

