% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbaseRest.R
\name{wormbaseRest}
\alias{wormbaseRest}
\title{WormBase RESTful API query}
\usage{
wormbaseRest(query, class, instance)
}
\description{
\url{http://www.wormbase.org/about/userguide/for_developers/API-REST}
}

