% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbaseRestRnaiSequence.R
\name{wormbaseRestRnaiSequence}
\alias{wormbaseRestRnaiSequence}
\title{WormBase RESTful RNAi sequence query.}
\usage{
wormbaseRestRnaiSequence(rnai)
}
\arguments{
\item{rnai}{WormBase RNAi identifier.}
}
\value{
tibble.
}
\description{
WormBase RESTful RNAi sequence query.
}
\examples{
wormbaseRestRnaiSequence("WBRNAi00003982")
}

