% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbaseRestRnaiTargets.R
\name{wormbaseRestRnaiTargets}
\alias{wormbaseRestRnaiTargets}
\title{WormBase RESTful RNAi targets query.}
\usage{
wormbaseRestRnaiTargets(wbrnai)
}
\arguments{
\item{wbrnai}{WormBase RNAi identifier vector.}
}
\value{
tibble.
}
\description{
WormBase RESTful RNAi targets query.
}
\examples{
wormbaseRestRnaiTargets("WBRNAi00031683")
}

