% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worminfo.R
\docType{package}
\name{worminfo-package}
\alias{worminfo}
\alias{worminfo-package}
\title{worminfo}
\description{
C. elegans genome annotations and RNAi clone mappings.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://steinbaugh.com/worminfo/}
  \item \url{https://github.com/steinbaugh/worminfo}
  \item Report bugs at \url{https://github.com/steinbaugh/worminfo/issues}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com}

}
