% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worminfo}
\alias{worminfo}
\title{\code{worminfo} Package Data}
\format{An object of class \code{list} of length 6.}
\usage{
worminfo
}
\description{
Gene and RNAi clone annotations.
}
\keyword{internal}
