% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahringer.R
\name{ahringer}
\alias{ahringer}
\title{Feeding RNAi Library clone matching}
\usage{
ahringer(id = NULL, wells = 384, format = "report")
}
\arguments{
\item{id}{Clone identifier}

\item{wells}{Library plate format (96, 384)}

\item{format}{Output type (report, simple)}
}
\value{
data.frame with metadata
}
\description{
Feeding RNAi Library clone matching
}
\examples{
ahringer("III-86@B01")
}

