% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-blastp.R
\name{blastp}
\alias{blastp}
\title{Best BLASTP hits}
\usage{
blastp(release = getOption(x = "wormbase.release", default = NULL))
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
WormBase release. If \code{NULL} (recommended), defaults to current production
release release available on the WormBase website. Legacy releases can be
specified as a character string (e.g. "WS267").}
}
\value{
\code{SplitDataFrameList}.
Split by \code{wormpep} column.
}
\description{
Best BLASTP hits
}
\note{
Updated 2022-06-08.
}
\examples{
x <- blastp()
print(x)
}
