% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastp.R
\name{blastp}
\alias{blastp}
\title{Best BLASTP hits}
\usage{
blastp(version = NULL, dir = ".")
}
\arguments{
\item{version}{\code{string} or \code{NULL}. WormBase version. If \code{NULL} (recommended),
defaults to current production release version available on the WormBase
website. Legacy versions can be specified as a \code{string} (e.g. "WS262").}

\item{dir}{\code{string}. Output directory.}
}
\value{
\code{tbl_df}. Grouped by \code{wormpep} column.
}
\description{
Best BLASTP hits
}
\examples{
x <- blastp()
glimpse(x)
}
