% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastp.R
\name{blastp}
\alias{blastp}
\title{Best BLASTP Hits}
\usage{
blastp(version = NULL, dir = ".")
}
\arguments{
\item{version}{WormBase version. If \code{NULL}, defaults to current production
release version available on the WormBase website. Legacy versions can
be specified as a string (e.g. "WS262").}

\item{dir}{Output directory.}
}
\value{
\link{tibble} grouped by wormpep.
}
\description{
Best BLASTP Hits
}
\examples{
blastp() \%>\% glimpse()
}
\seealso{
Other Annotation File Functions: \code{\link{annotationFile}},
  \code{\link{description}}, \code{\link{geneIDs}},
  \code{\link{geneOtherIDs}}, \code{\link{oligos}},
  \code{\link{orthologs}}
}
