% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-blastp.R
\name{blastp}
\alias{blastp}
\title{Best BLASTP hits}
\usage{
blastp(version = getOption(x = "wormbase.version", default = "WS269"))
}
\arguments{
\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS262").}
}
\value{
\code{tbl_df}. Grouped by \code{wormpep} column.
}
\description{
Best BLASTP hits
}
\examples{
x <- blastp()
glimpse(x)
}
