% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cherrypick.R
\name{cherrypick}
\alias{cherrypick}
\title{Cherrypick RNAi Clones by Keyword}
\usage{
cherrypick(identifier, format = "keyword")
}
\arguments{
\item{identifier}{Keyword identifier.}

\item{format}{Identifier format.}
}
\value{
RNAi clone \link{list} by gene.
}
\description{
Cherrypick RNAi Clones by Keyword
}
\examples{
cherrypick("unfolded protein response", format = "keyword") \%>\%
    glimpse()
}
