% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cherrypick.R
\name{cherrypick}
\alias{cherrypick}
\title{Cherrypick RNAi clones by keyword}
\usage{
cherrypick(keyword)
}
\arguments{
\item{keyword}{Keyword}
}
\value{
RNAi clone list by gene
}
\description{
Cherrypick RNAi clones by keyword
}
\examples{
cherrypick("transcription")
}

