% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cherrypick.R
\name{cherrypick}
\alias{cherrypick}
\title{Cherrypick RNAi clones by keyword}
\usage{
cherrypick(identifier, format = "keyword", ahringer384 = TRUE,
  ahringer96 = FALSE, orfeome96 = TRUE, plates = NULL)
}
\arguments{
\item{identifier}{Keyword identifier}

\item{format}{Identifier format}

\item{ahringer384}{Include Ahringer 384 well library}

\item{ahringer96}{Include Ahringer 96 well library}

\item{orfeome96}{Include ORFeome 96 well library}

\item{plates}{Reference character of plates to pick from}
}
\value{
RNAi clone list by gene
}
\description{
Cherrypick RNAi clones by keyword
}

