% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.R
\name{clone}
\alias{clone}
\title{Feeding RNAi Library clone matching}
\usage{
clone(id = NULL, library = "orfeome", wells = NULL, output = "simple")
}
\arguments{
\item{id}{Clone identifier}

\item{library}{Library type ("orfeome" or "ahringer")}

\item{wells}{Library plate format (96, 384)}

\item{output}{Output format (report, simple)}
}
\value{
tibble with \code{gene()} metadata
}
\description{
Feeding RNAi Library clone matching
}
\examples{
These should all match SBP-1:

ORFeome
clone("orfeome96-11010-G06", library = "orfeome")
clone("11010-G06", library = "orfeome")
clone("GHR-11010@G06", library = "orfeome")

Ahringer
clone("III-86B01", library = "ahringer", wells = 96)
clone("III-086-B01", library = "ahringer", wells = 96)
clone("086-B01", library = "ahringer", wells = 96)
clone("III-006-C01", library = "ahringer", wells = 384)
clone("III-6C1", library = "ahringer", wells = 384)
}

