% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.R
\name{clone}
\alias{clone}
\title{Feeding RNAi Library clone matching.}
\usage{
clone(identifier = NULL, format = "clone", library = "orfeome",
  wells = NULL)
}
\arguments{
\item{identifier}{Identifier.}

\item{format}{Identifier format - \code{clone} (e.g. GHR-11010@G06), \code{gene} (e.g. sbp-1), \code{genePair} (e.g. Y47D3B.7), \code{sequence} (e.g. Y47D3B.7))}

\item{format}{Identifier format ("cloneId" or "genePair")}

\item{library}{Library type ("orfeome" or "ahringer").}

\item{wells}{Library plate format (96, 384).}
}
\value{
tibble with \code{gene()} metadata.
}
\description{
Feeding RNAi Library clone matching.
}
\examples{
clone("GHR-11010@G06")
clone("086-B01", library = "ahringer", wells = 96)
clone("III-006-C01", library = "ahringer", wells = 384)
clone("Y47D3B.7", format = "genePair")
clone("sbp-1", format = "name")
}

