% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-currentRelease.R
\name{currentRelease}
\alias{currentRelease}
\title{Current WormBase release}
\usage{
currentRelease()
}
\value{
\code{character(1)}.
WormBase release version (e.g. WS271).
}
\description{
Current WormBase release
}
\note{
Updated 2019-07-27.
}
\examples{
## WormBase FTP server must be accessible.
tryCatch(
    expr = currentRelease(),
    error = function(e) e
)
}
