% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{dataRaw}
\alias{dataRaw}
\title{Source data-raw R script if necessary then load data binary}
\usage{
dataRaw(data)
}
\arguments{
\item{data}{Data filename}
}
\description{
Source data-raw R script if necessary then load data binary
}
\keyword{internal}

