% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-description.R
\name{description}
\alias{description}
\title{Gene functional descriptions}
\usage{
description(
  version = getOption(x = "wormbase.version", default = NULL),
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS267").}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:register]{bpparam}.
\item \link[BiocParallel:SerialParam-class]{SerialParam}.
\item \link[BiocParallel:MulticoreParam-class]{MulticoreParam}.
}}
}
\value{
\code{DataFrame}.
}
\description{
Gene functional descriptions
}
\note{
As of WS269 release, some non-N2 gene IDs are included in the flat
files available on the WormBase FTP server. These annotations are removed
from the return here, using grep matching to return only \code{WBGene} entries.

This file is malformed on the WormBase FTP server for WS270 and WS271
releases.

Updated 2021-02-17.
}
\examples{
## WormBase FTP server must be accessible.
tryCatch(
    expr = description(),
    error = function(e) e
)
}
