% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{description}
\alias{description}
\title{Gene functional descriptions}
\usage{
description(version = NULL, dir = ".", progress = FALSE)
}
\arguments{
\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS262").}

\item{dir}{\code{character(1)}.
Output directory.}

\item{progress}{\code{logical(1)}.
Show progress using pbapply package.}
}
\value{
\code{tbl_df}.
}
\description{
Gene functional descriptions
}
\note{
As of WS269 release, some non-N2 gene IDs are included in the flat
files available on the WormBase FTP server. These annotations are removed
from the return here, using grep matching to return only \code{WBGene} entries.
}
\examples{
x <- description()
glimpse(x)
}
