% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-description.R
\name{description}
\alias{description}
\title{Gene functional descriptions}
\usage{
description(release = getOption(x = "wormbase.release", default = NULL))
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
WormBase release. If \code{NULL} (recommended), defaults to current production
release release available on the WormBase website. Legacy releases can be
specified as a character string (e.g. "WS267").}
}
\value{
\code{DFrame}.
}
\description{
Gene functional descriptions
}
\note{
As of WS269 release, some non-N2 gene IDs are included in the flat
files available on the WormBase FTP server. These annotations are removed
from the return here, using grep matching to return only \code{WBGene} entries.

This file is malformed on the WormBase FTP server for WS270 and WS271
releases.

Updated 2023-09-25.
}
\examples{
x <- description()
print(x)
}
