% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-description.R
\name{description}
\alias{description}
\title{Gene functional descriptions}
\usage{
description(version = getOption(x = "wormbase.version", default = NULL),
  BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS262").}
}
\value{
\code{DataFrame}.
}
\description{
Gene functional descriptions
}
\note{
As of WS269 release, some non-N2 gene IDs are included in the flat
files available on the WormBase FTP server. These annotations are removed
from the return here, using grep matching to return only \code{WBGene} entries.

This file is currently malformed on the WormBase FTP server for WS270
and WS271 releases.

Updated 2019-08-28.
}
\examples{
## WormBase FTP server must be accessible.
## Currently failing for WS270, WS271 releases.
tryCatch(
    expr = description(version = "WS269"),
    error = function(e) e
)
}
