% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggnog.R
\name{eggnog}
\alias{eggnog}
\title{EggNOG annotations}
\usage{
eggnog(identifier)
}
\arguments{
\item{identifier}{EggNOG identifier}
}
\value{
tibble
}
\description{
EggNOG annotations
}
\examples{
c("ENOG410XPQV", "KOG0289") \%>\% eggnog
}

