% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggnog.R
\name{eggnog}
\alias{eggnog}
\title{EggNOG Annotations}
\usage{
eggnog(identifier)
}
\arguments{
\item{identifier}{EggNOG identifier.}
}
\value{
\link{tibble}.
}
\description{
EggNOG Annotations
}
\examples{
# daf-2
eggnog(c("KOG4258", "COG0515")) \%>\% glimpse()

# Multiple EGGNOG letters
eggnog("ENOG410IU5G") \%>\% glimpse()
}
