% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{ensemblBasic}
\alias{ensemblBasic}
\alias{ensemblGeneOntology}
\alias{ensemblInterpro}
\alias{ensemblPeptide}
\title{\href{http://www.ensembl.org}{Ensembl} utility functions}
\usage{
ensemblBasic(gene = NULL)

ensemblGeneOntology(gene = NULL, nest = TRUE)

ensemblInterpro(gene = NULL, nest = TRUE)

ensemblPeptide(peptide = NULL)
}
\arguments{
\item{gene}{Ensembl gene identifier.}

\item{nest}{Return as nested tibble.}

\item{peptide}{Ensembl peptide identifier.}
}
\value{
Ensembl metadata tibble.
}
\description{
\href{http://www.ensembl.org}{Ensembl} utility functions
}
\examples{
gene <- "WBGene00000001"
peptide <- "ENSP00000380252"
ensemblBasic(gene) \%>\% glimpse
ensemblGeneOntology(gene, nest = FALSE) \%>\% glimpse
ensemblInterpro(gene, nest = FALSE) \%>\% glimpse
ensemblPeptide(peptide) \%>\% glimpse
}
