% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{ensemblBasic}
\alias{ensemblBasic}
\alias{ensemblGeneOntology}
\alias{ensemblInterpro}
\alias{ensemblPeptide}
\title{Ensembl utility functions}
\usage{
ensemblBasic(gene = NULL)

ensemblGeneOntology(gene = NULL, nest = TRUE)

ensemblInterpro(gene = NULL, nest = TRUE)

ensemblPeptide(peptide = NULL)
}
\arguments{
\item{gene}{Ensembl gene identifier}

\item{nest}{Return as nested tibble}

\item{peptide}{Ensembl peptide identifier}
}
\value{
Ensembl metadata
}
\description{
Basic metadata

Gene Ontology terms

InterPro metadata

\emph{H. sapiens} annotations for WormBase BLASTP peptide hits
}
\examples{
ensemblBasic("WBGene00000001") \%>\% glimpse
ensemblGeneOntology("WBGene00000001", nest = FALSE) \%>\% glimpse
ensemblInterpro("WBGene00000001", nest = FALSE) \%>\% glimpse
ensemblPeptide("ENSP00000380252") \%>\% glimpse
}
