% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-externalIDs.R
\name{externalIDs}
\alias{externalIDs}
\title{External identifiers}
\usage{
externalIDs(genes, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{genes}{\code{character}.
Gene identifiers.}
}
\value{
\code{DataFrame}.
}
\description{
External identifiers
}
\note{
Updated 2019-08-29.
}
\examples{
## WormBase REST API must be accessible.
genes <- c("WBGene00000912", "WBGene00004804")
tryCatch(
    expr = externalIDs(genes),
    error = function(e) e
)
}
