% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-externalIDs.R
\name{externalIDs}
\alias{externalIDs}
\title{External identifiers}
\usage{
externalIDs(genes, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{genes}{\code{character}.
Gene identifiers (e.g. WBGene00004804).}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:bpparam]{bpparam}.
\item \link[BiocParallel:SerialParam]{SerialParam}.
\item \link[BiocParallel:MulticoreParam]{MulticoreParam}.
}}
}
\value{
\code{DataFrame}.
}
\description{
External identifiers
}
\note{
Updated 2020-01-03.
}
\examples{
## WormBase REST API must be accessible.
genes <- c("WBGene00000912", "WBGene00004804")
tryCatch(
    expr = externalIDs(genes),
    error = function(e) e
)
}
