% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-externalIDs.R
\name{externalIDs}
\alias{externalIDs}
\title{External identifiers}
\usage{
externalIDs(genes, progress = FALSE)
}
\arguments{
\item{genes}{\code{character}.
Gene identifiers.}

\item{progress}{\code{logical(1)}.
Show progress using pbapply package.}
}
\value{
\code{tbl_df}.
}
\description{
External identifiers
}
\examples{
x <- externalIDs(c("WBGene00000912", "WBGene00004804"))
glimpse(x)
}
