% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalIDs.R
\name{externalIDs}
\alias{externalIDs}
\title{External Identifiers}
\usage{
externalIDs(gene)
}
\arguments{
\item{gene}{Gene identifier.}
}
\value{
Gene \link{tibble}.
}
\description{
External Identifiers
}
\examples{
externalIDs(c("WBGene00000912", "WBGene00004804")) \%>\% glimpse()
}
\seealso{
Other REST API Functions: \code{\link{geneOntology}},
  \code{\link{rest}}
}
