% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalIDs.R
\name{externalIDs}
\alias{externalIDs}
\title{External Identifiers}
\usage{
externalIDs(genes)
}
\arguments{
\item{genes}{Gene identifiers.}
}
\value{
\code{tbl_df}.
}
\description{
External Identifiers
}
\examples{
invisible(capture.output(
    x <- externalIDs(c("WBGene00000912", "WBGene00004804"))
))
glimpse(x)
}
\seealso{
Other REST API Functions: \code{\link{geneOntology}}
}
