% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{gene}
\alias{gene}
\title{Gene annotations}
\usage{
gene(rowNames = NULL, id = "geneId", output = "report")
}
\arguments{
\item{rowNames}{Identifier vector}

\item{id}{Identifier type (geneID, orf, publicName)}

\item{output}{Output type (report, simple)}
}
\value{
metadata data.frame
}
\description{
Gene annotations
}

