% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{gene}
\alias{gene}
\title{Gene annotations.}
\usage{
gene(id = NULL, format = "geneId", select = "simple")
}
\arguments{
\item{id}{Identifier.}

\item{format}{Identifier type (geneID, orf, publicName).}

\item{select}{Columns to select (report, simple).}
}
\value{
tibble.
}
\description{
Gene annotations.
}
\examples{
gene(id = "WBGene00004804", format = "geneId")
gene(id = "T19E7.2", format = "orf")
gene(id = "skn-1", format = "publicName")
}

