% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{gene}
\alias{gene}
\title{Gene Matching}
\usage{
gene(identifier, format = "gene", select = NULL)
}
\arguments{
\item{identifier}{Identifier.}

\item{format}{Identifier type (\code{gene}, \code{sequence}, \code{name}, \code{class} or
\code{keyword}).}

\item{select}{Columns to select. Consult \code{vignette("gene")} for the list of
available parameters.}
}
\value{
\link{tibble}.
}
\description{
Gene Matching
}
\examples{
# Match by WormBase/Ensembl gene identifier (preferred)
gene("WBGene00004804", format = "gene") \%>\% glimpse()

# Match by gene sequence (aka ORF)
gene("T19E7.2", format = "sequence") \%>\% glimpse()

# Match by gene name (aka symbol)
gene("skn-1", format = "name") \%>\% glimpse()

# Match by gene class
gene("daf", format = "class") \%>\% glimpse()

# Match by keyword
gene("bzip", format = "keyword") \%>\% glimpse()

# Match by gene ID and select class
gene(
    "WBGene00004804",
    format = "gene",
    select = c("class", "status")) \%>\%
    glimpse()
}
