% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{gene}
\alias{gene}
\title{Gene mapping}
\usage{
gene(identifier, format = "gene", select = NULL)
}
\arguments{
\item{identifier}{Identifier}

\item{format}{Identifier type (\code{gene}, \code{name}, \code{sequence},
\code{class} or \code{keyword})}

\item{select}{Columns to select. Consult the \code{gene} vignette for
available parameters.}
}
\value{
tibble
}
\description{
Gene mapping
}
\examples{
gene("skn-1", format = "name")
gene("T19E7.2", format = "sequence")
gene("WBGene00004804", select = "descriptionConcise")
gene("daf", format = "class")
gene("bzip", format = "keyword")
}

