% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{gene}
\alias{gene}
\title{Gene annotations}
\usage{
gene(id = NULL, format = "geneId", output = "report")
}
\arguments{
\item{id}{Identifier}

\item{format}{Identifier type (geneID, orf, publicName)}

\item{output}{Output type (report, simple)}
}
\value{
metadata data.frame
}
\description{
Gene annotations
}
\examples{
gene()
gene(output = "simple")
gene(id = c("WBGene00004804", "WBGene00001752"), format = "geneId")
gene(id = c("T19E7.2", "K08F4.7"), format = "orf")
gene(id = c("skn-1", "gst-4"), format = "publicName")
}

