% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{gene}
\alias{gene}
\title{Gene mapping}
\usage{
gene(query, format = "gene", select = NULL, sort = NULL)
}
\arguments{
\item{query}{Identifier query}

\item{format}{Identifier type (\code{gene}, \code{name}, \code{sequence},
\code{class} or \code{keyword})}

\item{select}{Columns to select. Optionally, you can use \code{simple},
\code{report} or \code{NULL} declarations here.}

\item{sort}{Columns to use for sorting.}
}
\value{
tibble
}
\description{
Gene mapping
}
\examples{
gene("skn-1", format = "name")
gene("T19E7.2", format = "sequence")
gene("WBGene00004804", format = "gene", select = "report")
gene("daf", format = "class")
gene("bzip", format = "keyword")
}

