% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{gene}
\alias{gene}
\title{Gene annotations}
\usage{
gene(identifier, format = "gene", select = NULL)
}
\arguments{
\item{identifier}{Gene identifier}

\item{format}{Identifier type (\code{gene}, \code{name} or \code{sequence})}

\item{select}{Columns to select (e.g. \code{ncbi}).
Optionally, you can use \code{all} or \code{report} declarations here.}
}
\value{
tibble
}
\description{
Gene annotations
}
\examples{
gene("WBGene00004804", select = "ncbi")
gene("skn-1", format = "name")
gene("T19E7.2", format = "sequence")
}

