% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{gene}
\alias{gene}
\title{Gene annotations.}
\usage{
gene(identifier, format = "gene", select = NULL)
}
\arguments{
\item{identifier}{Gene identifier.}

\item{format}{Identifier type (gene, name, sequence).}

\item{select}{Columns to select (report, simple).}
}
\value{
tibble.
}
\description{
Gene annotations.
}
\examples{
gene("WBGene00004804")
gene("WBGene00004804", select = "ncbi")
gene("WBGene00004804", select = "all")
gene("skn-1", format = "name")
gene("T19E7.2", format = "sequence")
}

