% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExternal.R
\name{geneExternal}
\alias{geneExternal}
\title{Gene External Identifiers}
\usage{
geneExternal(gene)
}
\arguments{
\item{gene}{Gene identifier.}
}
\value{
Gene \link{tibble}.
}
\description{
Gene External Identifiers
}
\examples{
geneExternal(c("WBGene00000912", "WBGene00004804")) \%>\% glimpse()
}
\seealso{
Other REST API Functions: \code{\link{geneOntology}},
  \code{\link{rest}}
}
