% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExternal.R
\name{geneExternal}
\alias{geneExternal}
\title{WormBase RESTful gene external query}
\usage{
geneExternal(identifier)
}
\arguments{
\item{identifier}{Gene identifier}
}
\value{
JSON content tibble
}
\description{
WormBase RESTful gene external query
}
\examples{
geneExternal("WBGene00000001")
}

