% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExternal.R
\name{geneExternal}
\alias{geneExternal}
\title{WormBase RESTful Gene External Query}
\usage{
geneExternal(identifier)
}
\arguments{
\item{identifier}{Gene identifier.}
}
\value{
JSON content \link{tibble}.
}
\description{
WormBase RESTful Gene External Query
}
\examples{
geneExternal("WBGene00004804") \%>\% glimpse()
}
