% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneID_vec}
\alias{GeneID_vec}
\title{GeneID_vec}
\format{A character vector.}
\usage{
GeneID_vec
}
\description{
A list of all the current WormBase GeneIDs available. Useful for setting data
frames to the same number of rows.
}
\keyword{datasets}

