% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-geneIDs.R
\name{geneIDs}
\alias{geneIDs}
\title{Gene identifiers}
\usage{
geneIDs(version = getOption(x = "wormbase.version", default = NULL))
}
\arguments{
\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS262").}
}
\value{
\code{DataFrame}.
}
\description{
Gene identifiers
}
\note{
As of WS269 release, some non-N2 gene IDs are included in the flat
files available on the WormBase FTP server. These annotations are removed
from the return here, using grep matching to return only \code{WBGene} entries.

Updated 2019-08-28.
}
\examples{
## WormBase FTP server must be accessible.
tryCatch(
    expr = geneIDs(),
    error = function(e) e
)
}
