% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntology.R
\name{geneOntology}
\alias{geneOntology}
\title{Gene Ontology}
\usage{
geneOntology(genes, progress = TRUE)
}
\arguments{
\item{genes}{\code{character}. Gene identifiers.}

\item{progress}{\code{boolean}. Show progress using pbapply package.}
}
\value{
\code{tbl_df}.
}
\description{
Gene Ontology
}
\examples{
x <- geneOntology(
    genes = c("WBGene00000912", "WBGene00004804"),
    progress = FALSE
)
glimpse(x)
}
