% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-geneOntology.R
\name{geneOntology}
\alias{geneOntology}
\title{Gene Ontology terms}
\usage{
geneOntology(genes, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{genes}{\code{character}.
Gene identifiers.}
}
\value{
\code{DataFrame}.
}
\description{
Gene Ontology terms
}
\note{
Updated 2019-08-29.
}
\examples{
## WormBase REST API must be accessible.
genes <- c("WBGene00000912", "WBGene00004804")
tryCatch(
    expr = geneOntology(genes),
    error = function(e) e
)
}
