% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-geneOntology.R
\name{geneOntology}
\alias{geneOntology}
\title{Gene Ontology terms}
\usage{
geneOntology(genes, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{genes}{\code{character}.
Gene identifiers (e.g. WBGene00004804).}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:bpparam]{bpparam}.
\item \link[BiocParallel:SerialParam]{SerialParam}.
\item \link[BiocParallel:MulticoreParam]{MulticoreParam}.
}}
}
\value{
\code{DataFrame}.
}
\description{
Gene Ontology terms
}
\note{
Updated 2019-08-29.
}
\examples{
## WormBase REST API must be accessible.
genes <- c("WBGene00000912", "WBGene00004804")
tryCatch(
    expr = geneOntology(genes),
    error = function(e) e
)
}
