% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntology.R
\name{geneOntology}
\alias{geneOntology}
\title{WormBase RESTful RNAi Gene Ontology Query}
\usage{
geneOntology(identifier)
}
\arguments{
\item{identifier}{Gene identifier.}
}
\value{
JSON content \link{tibble}.
}
\description{
WormBase RESTful RNAi Gene Ontology Query
}
\examples{
geneOntology("WBGene00000001") \%>\% glimpse()
}
