% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-geneOntology.R
\name{geneOntology}
\alias{geneOntology}
\title{Gene Ontology terms}
\usage{
geneOntology(genes, progress = FALSE)
}
\arguments{
\item{genes}{\code{character}.
Gene identifiers.}

\item{progress}{\code{logical(1)}.
Show progress using pbapply package.}
}
\value{
\code{DataFrame}.
}
\description{
Gene Ontology terms
}
\note{
Updated 2019-08-28.
}
\examples{
## WormBase REST API must be accessible.
tryCatch(
    expr = geneOntology(c("WBGene00000912", "WBGene00004804")),
    error = function(e) e
)
}
