% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntology.R
\name{geneOntology}
\alias{geneOntology}
\title{Gene Ontology}
\usage{
geneOntology(genes)
}
\arguments{
\item{genes}{Gene identifiers.}
}
\value{
\code{tbl_df}.
}
\description{
Gene Ontology
}
\examples{
invisible(capture.output(
    x <- geneOntology(c("WBGene00000912", "WBGene00004804"))
))
glimpse(x)
}
\seealso{
Other REST API Functions: \code{\link{externalIDs}}
}
\concept{REST API Functions}
