% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-geneOntology.R
\name{geneOntology}
\alias{geneOntology}
\title{Gene Ontology terms}
\usage{
geneOntology(genes)
}
\arguments{
\item{genes}{\code{character}.
Gene identifiers (e.g. WBGene00004804).}
}
\value{
\code{List}.
}
\description{
Gene Ontology terms
}
\note{
Updated 2021-02-18.
}
\examples{
genes <- c("WBGene00000912", "WBGene00004804")
x <- geneOntology(genes)
print(x)
}
