% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntology.R
\name{geneOntology}
\alias{geneOntology}
\title{WormBase RESTful RNAi gene ontology query}
\usage{
geneOntology(identifier)
}
\arguments{
\item{identifier}{Gene identifier}
}
\value{
JSON content tibble
}
\description{
WormBase RESTful RNAi gene ontology query
}

