% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOntology.R
\name{geneOntology}
\alias{geneOntology}
\title{Gene Ontology}
\usage{
geneOntology(gene)
}
\arguments{
\item{gene}{Gene identifier.}
}
\value{
Gene \link{tibble}.
}
\description{
Gene Ontology
}
\examples{
geneOntology(c("WBGene00000912", "WBGene00004804")) \%>\% glimpse()
}
\seealso{
Other REST API Functions: \code{\link{externalIDs}},
  \code{\link{rest}}
}
