% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOtherIDs.R
\name{geneOtherIDs}
\alias{geneOtherIDs}
\title{Other Gene Identifiers}
\usage{
geneOtherIDs(version = NULL, dir = ".")
}
\arguments{
\item{version}{\code{string} or \code{NULL}. WormBase version. If \code{NULL} (recommended),
defaults to current production release version available on the WormBase
website. Legacy versions can be specified as a \code{string} (e.g. "WS262").}

\item{dir}{\code{string}. Output directory.}
}
\value{
\code{tbl_df}.
}
\description{
Other Gene Identifiers
}
\examples{
x <- geneOtherIDs()
glimpse(x)
}
