% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOtherIDs.R
\name{geneOtherIDs}
\alias{geneOtherIDs}
\title{Other Gene Identifiers}
\usage{
geneOtherIDs(version = NULL, dir = ".")
}
\arguments{
\item{version}{WormBase version. If \code{NULL}, defaults to current production
release version available on the WormBase website. Legacy versions can
be specified as a string (e.g. "WS262").}

\item{dir}{Output directory.}
}
\value{
\code{tbl_df}.
}
\description{
Other Gene Identifiers
}
\examples{
geneOtherIDs() \%>\% glimpse()
}
\seealso{
Other FTP File Functions: \code{\link{blastp}},
  \code{\link{description}}, \code{\link{geneIDs}},
  \code{\link{oligos}}, \code{\link{orthologs}},
  \code{\link{peptides}}, \code{\link{rnaiPhenotypes}}
}
