% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneReport.R
\name{geneReport}
\alias{geneReport}
\title{Gene list report}
\usage{
geneReport(query)
}
\arguments{
\item{query}{Gene identifier}
}
\value{
tibble
}
\description{
Gene list report
}
\examples{
geneReport("WBGene00000001")
}

