% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneReport.R
\name{geneReport}
\alias{geneReport}
\title{Gene list report}
\usage{
geneReport(identifier, format = "gene")
}
\arguments{
\item{identifier}{Gene identifier}

\item{format}{Identifier format}
}
\value{
tibble
}
\description{
Gene list report
}
\author{
Michael Steinbaugh
}
