% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneReport.R
\name{geneReport}
\alias{geneReport}
\title{Gene List Report}
\usage{
geneReport(identifier, format = "gene")
}
\arguments{
\item{identifier}{Gene identifier.}

\item{format}{Identifier format.}
}
\value{
\link{tibble}.
}
\description{
Gene List Report
}
\examples{
geneReport("daf-2", format = "name") \%>\%
    glimpse()
}
