% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical2wbrnai.R
\name{historical2wbrnai}
\alias{historical2wbrnai}
\title{Convert historical RNAi identifier to WBRNAi with WormBase RESTful API.}
\usage{
historical2wbrnai(historical)
}
\arguments{
\item{historical}{WormBase historical RNAi experiment vector.}
}
\value{
tibble.
}
\description{
Convert historical RNAi identifier to WBRNAi with WormBase RESTful API.
}
\examples{
historical2wbrnai("JA:K10E9.1")
}

