% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{metadata}
\format{A data frame with 44 variables:
  \describe{
  \item{\code{geneId}}{a character vector}
  \item{\code{orf}}{a character vector}
  \item{\code{publicName}}{a character vector}
  \item{\code{wormbaseStatus}}{a character vector}
  \item{\code{wormbaseGeneOtherIds}}{a character vector}
  \item{\code{wormbaseConciseDescription}}{a character vector}
  \item{\code{wormbaseProvisionalDescription}}{a character vector}
  \item{\code{wormbaseDetailedDescription}}{a character vector}
  \item{\code{wormbaseAutomatedDescription}}{a character vector}
  \item{\code{wormbaseGeneClassDescription}}{a character vector}
  \item{\code{wormbaseRnaiPhenotypes}}{a character vector}
  \item{\code{wormbaseOrthologsHsapiensId}}{a character vector}
  \item{\code{wormbaseOrthologsHsapiensName}}{a character vector}
  \item{\code{wormbaseBlastpWormpepId}}{a character vector}
  \item{\code{wormbaseBlastpEnsemblPeptideId}}{a numeric vector}
  \item{\code{wormbaseBlastpEValue}}{a character vector}
  \item{\code{wormbaseBlastpEnsemblGeneId}}{a character vector}
  \item{\code{wormbaseBlastpEnsemblGeneName}}{a character vector}
  \item{\code{wormbaseBlastpEnsemblDescription}}{a character vector}
  \item{\code{ensemblGeneBiotype}}{a character vector}
  \item{\code{ensemblChromosomeName}}{a numeric vector}
  \item{\code{ensemblStartPosition}}{a numeric vector}
  \item{\code{ensemblEndPosition}}{a numeric vector}
  \item{\code{ensemblStrand}}{a character vector}
  \item{\code{ensemblDescription}}{a character vector}
  \item{\code{ensemblEntrezGeneId}}{a character vector}
  \item{\code{ensemblRefseqMrna}}{a character vector}
  \item{\code{ensemblRefseqNcrna}}{a character vector}
  \item{\code{ensemblUniprotSptrembl}}{a character vector}
  \item{\code{ensemblUniprotSwissprot}}{a character vector}
  \item{\code{ensemblGeneOntologyId}}{a character vector}
  \item{\code{ensemblGeneOntologyName}}{a character vector}
  \item{\code{ensemblInterpro}}{a character vector}
  \item{\code{ensemblInterproShortDescription}}{a character vector}
  \item{\code{ensemblInterproDescription}}{a character vector}
  \item{\code{pantherUniprotKb}}{a character vector}
  \item{\code{pantherSubfamilyId}}{a character vector}
  \item{\code{pantherFamilyName}}{a character vector}
  \item{\code{pantherSubfamilyName}}{a character vector}
  \item{\code{pantherGeneOntologyMolecularFunction}}{a character vector}
  \item{\code{pantherGeneOntologyBiologicalProcess}}{a character vector}
  \item{\code{pantherGeneOntologyCellularComponent}}{a character vector}
  \item{\code{pantherClass}}{a character vector}
  \item{\code{pantherPathway}}{a character vector}
  }}
\source{
WormBase (\url{http://www.wormbase.org}),
Biomart (\url{http://www.biomart.org}),
InterPro (\url{https://www.ebi.ac.uk/interpro}),
PANTHER (\url{http://pantherdb.org})
}
\usage{
metadata
}
\description{
C. elegans metadata pulled from WormBase, Ensembl, and PANTHER.
}
\keyword{datasets}

