% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{metadata}
\format{A data frame with 44 variables:
  \describe{
  \item{\code{GeneID}}{a character vector}
  \item{\code{public.name}}{a character vector}
  \item{\code{ORF}}{a character vector}
  \item{\code{wormbase.status}}{a character vector}
  \item{\code{gene.other.ids}}{a character vector}
  \item{\code{description.gene.class}}{a character vector}
  \item{\code{description.concise}}{a character vector}
  \item{\code{description.provisional}}{a character vector}
  \item{\code{description.automated}}{a character vector}
  \item{\code{rnai.phenotypes}}{a character vector}
  \item{\code{orthologs.hsapiens.homolog.wormbase.id}}{a character vector}
  \item{\code{orthologs.hsapiens.homolog.wormbase.name}}{a character vector}
  \item{\code{blastp.wormbase.peptide.id}}{a character vector}
  \item{\code{blastp.ensembl.peptide.id}}{a character vector}
  \item{\code{blastp.e.val}}{a numeric vector}
  \item{\code{blastp.ensembl.gene.id}}{a character vector}
  \item{\code{blastp.external.gene.name}}{a character vector}
  \item{\code{blastp.description}}{a character vector}
  \item{\code{biomart.gene.biotype}}{a character vector}
  \item{\code{biomart.chromosome.name}}{a character vector}
  \item{\code{biomart.start.position}}{a numeric vector}
  \item{\code{biomart.end.position}}{a numeric vector}
  \item{\code{biomart.strand}}{a numeric vector}
  \item{\code{biomart.ensembl.description}}{a character vector}
  \item{\code{biomart.entrezgene}}{a character vector}
  \item{\code{biomart.refseq.mrna}}{a character vector}
  \item{\code{biomart.refseq.ncrna}}{a character vector}
  \item{\code{biomart.uniprot.sptrembl}}{a character vector}
  \item{\code{biomart.uniprot.swissprot}}{a character vector}
  \item{\code{biomart.hsapiens.homolog.ensembl.gene}}{a character vector}
  \item{\code{biomart.ensembl.go.id}}{a character vector}
  \item{\code{biomart.ensembl.go.names}}{a character vector}
  \item{\code{biomart.interpro}}{a character vector}
  \item{\code{biomart.interpro.short.description}}{a character vector}
  \item{\code{biomart.interpro.description}}{a character vector}
  \item{\code{panther.uniprot.kb}}{a character vector}
  \item{\code{panther.sf.id}}{a character vector}
  \item{\code{panther.family.name}}{a character vector}
  \item{\code{panther.subfamily.name}}{a character vector}
  \item{\code{panther.go.mf}}{a character vector}
  \item{\code{panther.go.bp}}{a character vector}
  \item{\code{panther.go.cc}}{a character vector}
  \item{\code{panther.pc}}{a character vector}
  \item{\code{panther.pathway}}{a character vector}
  }}
\source{
WormBase (\url{http://www.wormbase.org}),
Biomart (\url{http://www.biomart.org}),
InterPro (\url{https://www.ebi.ac.uk/interpro}),
PANTHER (\url{http://pantherdb.org})
}
\usage{
metadata
}
\description{
C. elegans metadata pulled from WormBase, Ensembl, and PANTHER.
}
\keyword{datasets}

