% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_ORF}
\alias{metadata_ORF}
\title{metadata_ORF}
\format{A data frame with 4 variables: \code{GeneID}, \code{ORF},
\code{public.name}, \code{gene.other.ids}}
\usage{
metadata_ORF
}
\description{
A dataset containing all current WormBase GeneIDs, public names, open reading
frames (ORFs), and previously used IDs. This dataset is filtered to have
unique ORF matches.
}
\keyword{datasets}

