% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microplate.R
\name{microplate}
\alias{microplate}
\title{Generate well identifiers for a microplate}
\usage{
microplate(plate = 1, well = 96, prefix = NULL)
}
\arguments{
\item{plate}{Number of plates}

\item{well}{Number of wells (\code{96}, \code{384})}

\item{prefix}{Plate name prefix}
}
\description{
Generate well identifiers for a microplate
}

