% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clones.R
\name{minimalClone}
\alias{minimalClone}
\title{Minimal RNAi Clones}
\usage{
minimalClone(identifier)
}
\arguments{
\item{identifier}{Clone identifier.}
}
\value{
Character vector.
}
\description{
Strip out unnecessary plate and well separators. Also remove zero padding
for well numbers, for fast and consistent clone matching.
}
\keyword{internal}
