% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oligos.R
\name{oligos}
\alias{oligos}
\title{PCR Oligo Sequences}
\usage{
oligos(version = NULL, dir = ".")
}
\arguments{
\item{version}{WormBase version. If \code{NULL}, defaults to current production
release version available on the WormBase website. Legacy versions can
be specified as a string (e.g. "WS262").}

\item{dir}{Output directory.}
}
\value{
\link{tibble} grouped by gene.
}
\description{
PCR Oligo Sequences
}
\examples{
oligos() \%>\% glimpse()
}
\seealso{
Other Annotation File Functions: \code{\link{annotationFile}},
  \code{\link{blastp}}, \code{\link{description}},
  \code{\link{geneIDs}}, \code{\link{geneOtherIDs}},
  \code{\link{orthologs}}
}
