% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oligos.R
\name{oligos}
\alias{oligos}
\title{PCR oligo sequences}
\usage{
oligos(version = NULL)
}
\arguments{
\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS262").}
}
\value{
\code{tbl_df}.
}
\description{
PCR oligo sequences
}
\examples{
x <- oligos()
glimpse(x)
}
