% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-oligos.R
\name{oligos}
\alias{oligos}
\title{PCR oligo sequences}
\usage{
oligos(release = getOption(x = "wormbase.release", default = NULL))
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
WormBase release. If \code{NULL} (recommended), defaults to current production
release release available on the WormBase website. Legacy releases can be
specified as a character string (e.g. "WS267").}
}
\value{
\code{CharacterList}.
}
\description{
PCR oligo sequences
}
\note{
Updated 2023-09-25.
}
\examples{
x <- oligos()
print(x)
}
