% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-orthologs.R
\name{orthologs}
\alias{orthologs}
\title{Orthologs}
\usage{
orthologs(release = getOption(x = "wormbase.release", default = NULL))
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
WormBase release. If \code{NULL} (recommended), defaults to current production
release release available on the WormBase website. Legacy releases can be
specified as a character string (e.g. "WS267").}
}
\value{
\code{CharacterList}.
}
\description{
Orthologs
}
\note{
Updated 2021-02-18.
}
\examples{
x <- orthologs()
print(x)
}
