% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-orthologs.R
\name{orthologs}
\alias{orthologs}
\title{Orthologs}
\usage{
orthologs(version = getOption(x = "wormbase.version", default = NULL),
  BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS262").}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:bpparam]{bpparam}.
\item \link[BiocParallel:SerialParam]{SerialParam}.
\item \link[BiocParallel:MulticoreParam]{MulticoreParam}.
}}
}
\value{
\code{DataFrame}.
}
\description{
Orthologs
}
\note{
Updated 2019-08-28.
}
\examples{
## WormBase FTP server must be accessible.
tryCatch(
    expr = orthologs(),
    error = function(e) e
)
}
