% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthologs.R
\name{orthologs}
\alias{orthologs}
\title{Orthologs}
\usage{
orthologs(version = NULL, dir = ".")
}
\arguments{
\item{version}{WormBase version. If \code{NULL}, defaults to current production
release version available on the WormBase website. Legacy versions can
be specified as a string (e.g. "WS262").}

\item{dir}{Output directory.}
}
\value{
\code{tbl_df}.
}
\description{
Orthologs
}
\examples{
invisible(capture.output(
    x <- orthologs()
))
glimpse(x)
}
\seealso{
Other FTP File Functions: \code{\link{blastp}},
  \code{\link{description}}, \code{\link{geneIDs}},
  \code{\link{geneOtherIDs}}, \code{\link{oligos}},
  \code{\link{peptides}}, \code{\link{rnaiPhenotypes}}
}
