% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{dir}{\code{string}. Output directory.}

\item{file}{\code{string}. Annotation file name.}

\item{gene}{\code{string}. Gene identifier.}

\item{genes}{\code{character}. Gene identifiers.}

\item{progress}{\code{boolean}. Show progress using pbapply package.}

\item{version}{\code{string} or \code{NULL}. WormBase version. If \code{NULL} (recommended),
defaults to current production release version available on the WormBase
website. Legacy versions can be specified as a \code{string} (e.g. "WS262").}
}
\description{
Parameters
}
\keyword{internal}
