% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{file}{\code{character(1)}.
Annotation file name.}

\item{gene}{\code{character(1)}.
Gene identifier.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{progress}{\code{logical(1)}.
Show progress using pbapply package.}

\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS262").}
}
\description{
Parameters
}
\keyword{internal}
