% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{file}{\code{character(1)}.
Annotation file name.}

\item{genes}{\code{character}.
Gene identifiers (e.g. WBGene00004804).}

\item{release}{\code{character(1)} or \code{NULL}.
WormBase release. If \code{NULL} (recommended), defaults to current production
release release available on the WormBase website. Legacy releases can be
specified as a character string (e.g. "WS267").}
}
\description{
Parameters
}
\keyword{internal}
