% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-peptides.R
\name{peptides}
\alias{peptides}
\title{Peptides}
\usage{
peptides(version = getOption(x = "wormbase.version", default = NULL),
  BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS262").}
}
\value{
\code{SplitDataFrameList}.
Split by \code{geneID} column.
}
\description{
Peptides
}
\note{
Updated 2019-08-28.
}
\examples{
## WormBase FTP server must be accessible.
tryCatch(
    expr = peptides(),
    error = function(e) e
)
}
