% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-peptides.R
\name{peptides}
\alias{peptides}
\title{Peptides}
\usage{
peptides(
  version = getOption(x = "wormbase.version", default = NULL),
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS262").}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:register]{bpparam}.
\item \link[BiocParallel:SerialParam-class]{SerialParam}.
\item \link[BiocParallel:MulticoreParam-class]{MulticoreParam}.
}}
}
\value{
\code{SplitDataFrameList}.
Split by \code{geneID} column.
}
\description{
Peptides
}
\note{
Updated 2020-01-26.
}
\examples{
## WormBase FTP server must be accessible.
tryCatch(
    expr = peptides(),
    error = function(e) e
)
}
