% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp-peptides.R
\name{peptides}
\alias{peptides}
\title{Peptides}
\usage{
peptides(release = getOption(x = "wormbase.release", default = NULL))
}
\arguments{
\item{release}{\code{character(1)} or \code{NULL}.
WormBase release. If \code{NULL} (recommended), defaults to current production
release release available on the WormBase website. Legacy releases can be
specified as a character string (e.g. "WS267").}
}
\value{
\code{SplitDataFrameList}.
Split by \code{geneId} column.
}
\description{
Peptides
}
\note{
Updated 2022-06-08.
}
\examples{
x <- peptides()
print(x)
}
