% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptides.R
\name{peptides}
\alias{peptides}
\title{Peptides}
\usage{
peptides(version = NULL, dir = ".", progress = FALSE)
}
\arguments{
\item{version}{\code{character(1)} or \code{NULL}.
WormBase version. If \code{NULL} (recommended), defaults to current production
release version available on the WormBase website. Legacy versions can be
specified as a character string (e.g. "WS262").}

\item{dir}{\code{character(1)}.
Output directory.}

\item{progress}{\code{logical(1)}.
Show progress using pbapply package.}
}
\value{
\code{tbl_df}. Grouped by \code{gene} column.
}
\description{
Peptides
}
\examples{
x <- peptides()
glimpse(x)
}
