% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{markdownList}
\alias{reexports}
\alias{microplate}
\alias{reexports}
\alias{file_delete}
\alias{reexports}
\alias{file_exists}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{markdownList}}, \code{\link[basejump]{microplate}}}

  \item{fs}{\code{\link[fs]{file_delete}}, \code{\link[fs]{file_exists}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}, \code{\link[tibble]{tibble}}}
}}

