% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{rest}
\alias{rest}
\title{RESTful API Query}
\usage{
rest(query)
}
\value{
\link{list}.
}
\description{
RESTful API Query
}
\seealso{
Other REST API Functions: \code{\link{geneExternal}},
  \code{\link{geneOntology}}
}
