% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restGeneExternal.R
\name{restGeneExternal}
\alias{restGeneExternal}
\title{WormBase RESTful gene external query}
\usage{
restGeneExternal(identifier)
}
\arguments{
\item{identifier}{Gene identifier}
}
\value{
JSON content tibble
}
\description{
WormBase RESTful gene external query
}
\examples{
restGeneExternal("WBGene00000001")
}

