% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rnai}
\alias{rnai}
\title{Feeding RNAi Library clone matching}
\usage{
rnai(cloneId = NULL, lib = "orfeome", type = "simple")
}
\arguments{
\item{cloneId}{Clone identifier}

\item{lib}{Feeding library (orfeome, ahringer)}

\item{type}{Output type (simple, report)}
}
\value{
data.frame with metadata
}
\description{
Feeding RNAi Library clone matching
}

