% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnai.R
\name{rnai}
\alias{rnai}
\title{RNAi clone mapping}
\usage{
rnai(identifier, format = "clone", proteinCoding = TRUE)
}
\arguments{
\item{identifier}{Identifier.}

\item{format}{Identifier format (\code{clone}, \code{gene}, \code{genePair}, \code{name}, or
\code{sequence}).}

\item{proteinCoding}{Return protein coding matches (\code{TRUE}/\code{FALSE}).}
}
\value{
Tibble.
}
\description{
RNAi clone mapping
}
\examples{
rnai("ahringer384-III-6-C01")
rnai("ahringer96-86-B01")
rnai("GHR-11010@G06")
rnai("orfeome96-11010-G06")
rnai("sbp-1", format = "name")
rnai("WBGene00004735", format = "gene")
rnai("Y47D3B.7", format = "sequence")
rnai("Y53H1C.b", format = "genePair")
}
