% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnai.R
\name{rnai}
\alias{rnai}
\title{Feeding RNAi Library clone matching}
\usage{
rnai(id = NULL, library = "orfeome", format = "report")
}
\arguments{
\item{id}{Clone identifier}

\item{library}{Feeding library (ahringer, orfeome)}

\item{format}{Output type (report, simple)}
}
\value{
data.frame with metadata
}
\description{
Feeding RNAi Library clone matching
}
\examples{
rnai(library = "ahringer")
rnai(id = "III-86@B01", library = "ahringer")
rnai(library = "orfeome")
rnai(id = "11010@G06", library = "orfeome")
}

