% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnai.R
\name{rnai}
\alias{rnai}
\title{RNAi clone mapping}
\usage{
rnai(identifier, format = "clone")
}
\arguments{
\item{identifier}{Identifier}

\item{format}{Identifier format (\code{clone}, \code{gene}, \code{genePair},
\code{name}, or \code{sequence})}
}
\value{
tibble
}
\description{
RNAi clone mapping
}
\examples{
c("GHR-11010@G06", "orfeome96-11010-G06", "ahringer384-III-6-C01", "ahringer96-86-B01") \%>\% rnai
"sbp-1" \%>\% rnai(format = "name")
"WBGene00004735" \%>\% rnai(format = "gene")
"Y47D3B.7" \%>\% rnai(format = "sequence")
"Y47D3B.7" \%>\% rnai(format = "genePair")
}

