% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnai.R
\name{rnai}
\alias{rnai}
\title{RNAi clone mapping}
\usage{
rnai(identifier, format = "clone", library = "orfeome96")
}
\arguments{
\item{identifier}{Identifier}

\item{format}{Identifier format (\code{gene}, \code{historical}, \code{name},
\code{rnai} or \code{sequence})}

\item{library}{Library type (\code{ahringer96}, \code{ahringer384},
\code{cherrypick} or \code{orfeome96})}
}
\value{
tibble
}
\description{
RNAi clone mapping
}
\examples{
rnai("sbp-1", format = "name")
rnai("WBGene00004735", format = "gene")
rnai("Y47D3B.7", format = "sequence")
rnai("JA:Y47D3B.7", format = "historical")
rnai("WBRNAi00009345", format = "rnai")
rnai("III-6C01", library = "ahringer384")
rnai("86B01", library = "ahringer96")
rnai("GHR-11010@G06", library = "orfeome96")
rnai("tf_all-1E1", library = "cherrypick")
}

