% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnai.R
\name{rnai}
\alias{rnai}
\title{RNAi clone matching}
\usage{
rnai(identifier, format = "clone", library = "orfeome96")
}
\arguments{
\item{identifier}{Identifier}

\item{format}{Identifier format (\code{gene}, \code{historical}, \code{name},
\code{rnai} or \code{sequence})}

\item{library}{Library type (\code{ahringer96}, \code{ahringer96Historical},
\code{ahringer384}, \code{cherrypick} or \code{orfeome})}
}
\value{
tibble
}
\description{
RNAi clone matching
}
\examples{
rnai("GHR-11049@F12", library = "orfeome")
rnai("IV-2N18", library = "ahringer384")
rnai("99G09", library = "ahringer96")
rnai("99G09", library = "ahringer96Historical")
rnai("tf_all-3C06", library = "cherrypick")
rnai("MV_SV:mv_T19E7.2", format = "historical")
rnai("WBRNAi00009186", format = "rnai")
rnai("T19E7.2", format = "sequence")
rnai("skn-1", format = "name")
}

