% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaiPhenotypes.R
\name{rnaiPhenotypes}
\alias{rnaiPhenotypes}
\title{RNAi Phenotypes}
\usage{
rnaiPhenotypes(version = NULL, dir = ".")
}
\arguments{
\item{version}{WormBase version. If \code{NULL}, defaults to current production
release version available on the WormBase website. Legacy versions can
be specified as a string (e.g. "WS262").}

\item{dir}{Output directory.}
}
\value{
Gene \link{tibble}.
}
\description{
RNAi Phenotypes
}
\examples{
rnaiPhenotypes() \%>\% glimpse()
}
\seealso{
Other FTP File Functions: \code{\link{blastp}},
  \code{\link{peptides}}
}
