% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\docType{data}
\name{simpleCol}
\alias{simpleCol}
\title{Simple columns}
\format{An object of class \code{character} of length 3.}
\usage{
simpleCol
}
\arguments{
\item{simpleCol}{Simple columns}
}
\description{
Simple columns
}
\keyword{internal}

