% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{uniprot}
\alias{uniprot}
\title{UniProt Web Service Query}
\usage{
uniprot(identifier)
}
\arguments{
\item{identifier}{WormBase gene identifier.}
}
\value{
\link{tibble}.
}
\description{
UniProt Web Service Query
}
