% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worfdb.R
\name{worfdbHTML}
\alias{worfdbHTML}
\alias{worfdbData}
\title{Query WORFDB website for RNAi clone annotations}
\usage{
worfdbHTML(sequence)

worfdbData(worfdbHTML)
}
\arguments{
\item{sequence}{Sequence identifier.}

\item{worfdbHTML}{List of WORFDB HTML pages.}
}
\description{
Query WORFDB website for RNAi clone annotations
}
\keyword{internal}
