% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worfdb.R
\name{worfdb}
\alias{worfdb}
\alias{worfdbHTML}
\alias{worfdbData}
\title{Query WORFDB Website for RNAi Clone Annotations}
\usage{
worfdbHTML(sequence)

worfdbData(worfdbHTML)
}
\arguments{
\item{sequence}{Sequence identifier.}

\item{worfdbHTML}{List of WORFDB HTML pages.}
}
\description{
Query WORFDB Website for RNAi Clone Annotations
}
\keyword{internal}
