% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbase-package.R
\docType{package}
\name{wormbase-package}
\alias{wormbase}
\alias{wormbase-package}
\title{wormbase}
\description{
\emph{C. elegans} genome annotations from WormBase.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://steinbaugh.com/wormbase/}
  \item \url{https://github.com/steinbaugh/wormbase}
  \item Report bugs at \url{https://github.com/steinbaugh/wormbase/issues}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com}

}
