% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbaseAnnotationFile.R
\name{wormbaseAnnotationFile}
\alias{wormbaseAnnotationFile}
\title{Download an annotation file from the WormBase FTP server}
\usage{
wormbaseAnnotationFile(file)
}
\arguments{
\item{file}{Annotation file name}
}
\value{
File path of download
}
\description{
Download an annotation file from the WormBase FTP server
}

