% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbaseAnnotationFile.R
\name{wormbaseAnnotationFile}
\alias{wormbaseAnnotationFile}
\title{WormBase Annotation File}
\usage{
wormbaseAnnotationFile(file)
}
\arguments{
\item{file}{Annotation file name.}
}
\value{
File path.
}
\description{
WormBase Annotation File
}
\keyword{internal}
