% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbaseFile.R
\name{wormbaseFile}
\alias{wormbaseFile}
\title{WormBase FTP server file download.}
\usage{
wormbaseFile(request)
}
\arguments{
\item{request}{WormBase FTP server file request.}
}
\value{
Local file path of downloaded file.
}
\description{
WormBase FTP server file download.
}

