% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbaseRest.R
\name{wormbaseRest}
\alias{wormbaseRest}
\title{WormBase RESTful API query}
\usage{
wormbaseRest(query, class, instance)
}
\arguments{
\item{query}{a WBGene ID (e.g. WBGene00006763)}

\item{class}{a class (e.g. gene)}

\item{instance}{an instance (e.g. concise_description)}
}
\description{
\url{http://www.wormbase.org/about/userguide/for_developers/API-REST}
}

