% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbaseRest.R
\name{wormbaseRest}
\alias{wormbaseRest}
\title{WormBase RESTful API query}
\usage{
wormbaseRest(query, class, instance)
}
\arguments{
\item{query}{A WormBase gene identifier (e.g. WBGene00000001).}

\item{class}{A class (e.g. gene).}

\item{instance}{An instance (e.g. concise_description).}
}
\description{
\url{http://www.wormbase.org/about/userguide/for_developers/API-REST}
}

