% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbase.R
\name{wormbaseRestGeneExternal}
\alias{wormbaseRestGeneExternal}
\title{WormBase RESTful gene identifier query.}
\usage{
wormbaseRestGeneExternal(gene)
}
\arguments{
\item{gene}{WormBase gene identifier.}
}
\value{
tibble.
}
\description{
WormBase RESTful gene identifier query.
}
\examples{
wormbaseRestGeneExternal("WBGene00000001")
}

