% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormbaseRestGeneExternal.R
\name{wormbaseRestGeneExternal}
\alias{wormbaseRestGeneExternal}
\title{Get Entrez identifier from WormBase Gene identifier.}
\usage{
wormbaseRestGeneExternal(gene)
}
\arguments{
\item{gene}{WormBase gene identifier.}
}
\value{
tibble.
}
\description{
Get Entrez identifier from WormBase Gene identifier.
}
\examples{
wormbaseRestGeneExternal("WBGene00000001")
}

